Fix drag reorder between DataViews
DataViews involved in a drag and drop operation are now handled in a central location (the frontscript). Messages are dispatched and affected views are restored to pre-drag state from the frontscript. dragEnd is responsible for triggering the cleanup handler in the frontscript and all are views are restored at once.

Note that drag and drop from a source outside of the engine is not supported. There is no message sent if a drag operation that started outside of the engine is canceled. Thus there is no way to reliably restore the DataView to pre-drag state.