If there are no records in the view then check the `dvAcceptsDrop` property of the view
For the view the `dvAcceptsDrop` property is a boolean specifying whether or not the empty view can accept the drop. Note that for rows `dvAcceptsDrop` returns an array of booleans. The keys are "above" and "below". Because "above" and "below" aren't applicable to the view as a whole a single boolean is returned.