Reset control height in no cache no fixed row height scenario
If cache is set to none and fixed row height is false and RenderRows is called then the cached heights of rows that should be rendered but which don’t currently have controls neeeds to be reset. This forces the control height to be recalculated the next time it is scrolled into view.