Performance Optimizations (#9)
* Use `control of id` references

The fastest reference to a control is `control id <id> of this stack`. It is constant time as the internal MCControl pointers are cached against the integer id in the stack object.

* Only update backgroundColor if it is changing

This will improve speed with acceleratedRendering when `container` is added for groups. If something doesn’t need to have a visual property updated then don’t do it.

* Only set the dvHilite custom property if it is changing

* Fix control existence check

* Early exit from vscroll if vscroll isn't changing