Do not automatically clear out unlinked cached controls
The DataView needs to support the notion of a row control being cached but hidden. For example, the DataView Tree will hide controls when toggling a node. If caching is turned on then those rows should remain in memory until they are actually deleted from the tree.

The cache can be modified or purged using `DeleteRowControlFromCache` and `ResetView`.