Do not call ResetView when setting the dvData
If a row template control in the DataView triggers a call that sets the dvData then an error will occur because that control cannot be deleted. Setting the dvData should just work.

Calling ResetView was done so that the developer would always see the latest template. If the developer needs to refresh the DataView template controls then they can explicitly call ResetView.