Add viewProp[“top padding template”] property
This property allows the developer to insert a group that will be displayed in the top padding are above the first row. Assign the property to the long id of a group. The group will be copied into the DataView when `RenderView` or `RefreshViewRows` is called.

Note that this control is not associated with any data in the DataView. It is not counted in the `number of rows` property. It will be sent a `FillInData` message with no parameters. When the DataView is resized it will receive the `resizeControl` message.