Initialize topleft of new controls to topleft of dvLIst
Doing so allows a developer to use the left of the DataView as a root value when determining positioning in the row templates. For example, you might describe CSS-like properties that row template behaviors can access for control positioning. If you can’t count on the control left being the left of dvList then you would need to pass item 1 of the pControlRect param that is passed to LayoutControl each time. That should not be necessary.