Fix potential eternal loop of doom
Observed behavior:
Imagine a row control that has an icon which is shown on MouseEnterRowControl and hidden on MouseLeaveRowControl. While visible, the user can click on it to show a popup menu. If the user moves the mouse around to other controls in the row control while the popup menu is visible, and then dismisses the menu while the mouse is outside of the DataView (by click or using escape key), an endless loop of mouseLeave/mouseEnter messages will be generated, triggered by the icon being shown.

Sending the mouseLeave message makes sense. mouseEnter does not make sense, however, as the mouse isn’t inside of the row template. Perhaps the engine stores some pointer to a control that isn’t being updated in this scenario?