Ensure initialization and mobile scroller
Due to `lockMessages` there is no guarantee that `newGroup` or `preOpenControl` will be called. Initialization is now triggered in `preOpenControl`, `openControl`, `_renderView` and `ResetView`. This covers all potential entry points.

In addition, the mobile scroller is not created in the initialization handler so that it will be created under all circumstances.