Use `effective styledText` to look for changes in a field
There are keys such as `spaceAbove` that can appear in the `styledText` keys that don’t actually change the appearance of the text but can trigger a false positive. Using the `effective styledText` ensures that all keys are always present so the diff is only being done on values that have actually changed.