Maintain field editor in proper row when rendering view and controls aren’t cached
Calling `RefreshViewRows` while a field editor is open and the `cache` setting is `none` could cause the field editor to appear in a different row as row controls were reassigned. This commit reassociates the field editor with the correct row when moving things around.

As part of this fix the sViewPropsA["fld_editor"]["control"] variable was changed to use the short id as that will not change as the field is relayered within the DataView.