Don’t restore the selectedChunk unnecessarily
The selectedChunk is not lost when removing focus during the scroll. A negative side effect is that restoring it can alter engine behavior. For example, if the user clicks and drags to select text in a field in the DataView, scrolls the DataView without releasing the mouse, and then continues to drag to select text, the drag selection no longer functions.