Add script for loading the DataView in non-Levure apps
`start using stack "DataView Assets Loader"` will load all of the necessary files for using DataViews whether running in the IDE or a standalone. When run in the IDE the files in the `ide` folder will be loaded. They help with creating new DataViews.

The IDE library will now prompt the user for a folder to save the template in if the Levure framework is not available.